Move IRMappingGroup to its own file.
